feat(*) Annotation logic moved to `ImageAnnotationController. Large commit!`
Created a master abstract class `Annotation` which is inherited from by `TextAnnotation` and `ShapeAnnotation` so that we can create a single `List<Annotation>`. This will allow us to properly manage undo and redo.

Added an option to pass in initial `color`, `strokeWidth` and `fontSize` which is stored and managed in the controller.

Moved `drawShape` to `paint_boundary_widget.dart` and replaced some attributes and passed in the controller instead.

Removed the old `text_annotation.dart` and moved the class into `annotation_models.dart`.

Removed `annotation_action.dart` as it is no longer a necessary part of the controller since all annotation logic has been moved INTO the controller. This also means no need for a callback action within the controller.